+++ src/nn/init.h
@@ -2,14 +2,14 @@
 #define INITIALIZER_H
 
 struct initializer_s;
-typedef void (*initializer_f)(const struct initializer_s* conf, long N, const long dims[N], _Complex float* weights);
+typedef void (*initializer_f)(const struct initializer_s* conf, int N, const long dims[N], _Complex float* weights);
 
-extern void initializer_apply(const struct initializer_s* conf, long N, const long dims[N], _Complex float* weights);
+extern void initializer_apply(const struct initializer_s* conf, int N, const long dims[N], _Complex float* weights);
 extern void initializer_free(const struct initializer_s* conf);
 extern const struct initializer_s* initializer_clone(const struct initializer_s* x);
 
-extern const struct initializer_s* init_reshape_create(unsigned int N, const long dims[N], const struct initializer_s* init);
-extern const struct initializer_s* init_stack_create(unsigned int N, int stack_dim, const long dimsa[N], const struct initializer_s* inita, const long dimsb[N], const struct initializer_s* initb);
+extern const struct initializer_s* init_reshape_create(int N, const long dims[N], const struct initializer_s* init);
+extern const struct initializer_s* init_stack_create(int N, int stack_dim, const long dimsa[N], const struct initializer_s* inita, const long dimsb[N], const struct initializer_s* initb);
 extern const struct initializer_s* init_dup_create(const struct initializer_s* inita, const struct initializer_s* initb);
 
 extern unsigned long in_flag_conv(_Bool c1);
@@ -26,6 +26,6 @@ extern const struct initializer_s* init_array_create(int N, const long dims[N], 
 extern const struct initializer_s* init_std_normal_create(_Bool real, float scale, float mean);
 extern const struct initializer_s* init_uniform_create(_Bool real, float scale, float mean);
 
-extern const struct initializer_s* init_linspace_create(unsigned int dim, _Complex float min_val, _Complex float max_val, _Bool max_inc);
+extern const struct initializer_s* init_linspace_create(int dim, _Complex float min_val, _Complex float max_val, _Bool max_inc);
 
 #endif
